requirements/base.txt
amqp==5.2.0
anyio==4.4.0
APScheduler==3.6.3
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==22.10.0
billiard==4.2.0
cachetools==4.2.2
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
cfgv==3.4.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
decorator==5.1.1
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
environ==1.0
filelock==3.15.4
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
identify==2.6.0
idna==3.7
incremental==24.7.2
kombu==5.4.0
msgpack==1.0.8
nodeenv==1.9.1
pillow==10.4.0
platformdirs==4.2.2
pre-commit==3.8.0
prompt_toolkit==3.0.47
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-environ==0.4.54
python-telegram-bot==13.7
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
service-identity==24.1.0
setuptools==72.2.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tornado==6.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.6
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
zope.interface==7.0.1
